+++ tests/examples/vite-react.example.js
@@ -59,12 +59,21 @@ root.render(
     '/package.json': JSON.stringify({
       scripts: { dev: 'vite' },
       dependencies: { react: '^18.2.0', 'react-dom': '^18.2.0' },
-      devDependencies: { vite: '^4.0.0', 'esbuild-wasm': '0.15.12' },
+      devDependencies: { vite: '^4.0.0', '@vitejs/plugin-react': '3.1.0', 'esbuild-wasm': '0.15.12' },
     }),
+    'vite.config.js': `import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+});`,
   });
 
   const shellProcess = emulator.shell.create();
-  const shellInfo = await shellProcess.runCommand('vite', ['dev', '--port', '3000']);
+  const shellInfo = await shellProcess.runCommand('vite', [], {
+    env: undefined,
+  });
   shellProcess.on('exit', (...data) => console.error('Process exited:', ...data));
   return shellInfo;
 };